refactor processor
